refactor(common-domain): import 수정
